Moved utf8bytes to file
